changed to functional component
